Se modifica el pom.xml agregando dependencia de jUnit y se cambian las pruebas para que fallen.

+++ src/com/sura/ensayandojenkis/domain/arithmeticfunctions/ArithmeticFunctions.java
@@ -19,7 +19,7 @@ public class ArithmeticFunctions {
 		return term1*term2;
 	}
 
-	public double division(int term1, int term2) {
+	public double division(Double term1, Double term2) {
 		return (term1/term2);
 	}
